% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{hessian}
\alias{hessian}
\title{Compute the Hessian-function of a function.}
\usage{
hessian(f, use_names = FALSE, ...)
}
\arguments{
\item{f}{A function}

\item{use_names}{Should the gradient add variable names to the output of the function?}

\item{vars}{The variables to compute the derivatives of. If NULL
the gradient will be computed for all formal arguments
of f.}
}
\value{
A function that computes the gradient of f at any point.
}
\description{
Creates a function that computes the second-order derivatives of a function with respect to each pair of parameters
and return a vector of these.
}
