% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacobian.R
\name{jacobian}
\alias{jacobian}
\title{Compute the gradient-function of a function.}
\usage{
jacobian(f, var)
}
\arguments{
\item{f}{A function}

\item{var}{The variable to compute the derivative of.}
}
\value{
A function that computes the jacobian of f at any point.
}
\description{
Creates a function that computes the jacobian matrix of a function with respect to one parameter
and returns a matrix of these.
}
